Time: 4 ms (85.83%), Space: 42.7 MB (60.75%) - LeetHub

+++ 1005-maximize-sum-of-array-after-k-negations/1005-maximize-sum-of-array-after-k-negations.java
@@ -1,9 +1,9 @@
 class Solution {
     public int largestSumAfterKNegations(int[] arr, int k) {
         Arrays.sort(arr);
-        long sum = 0;
-        long maxOfNegative = Integer.MAX_VALUE;
-        long minOfPositive = Integer.MAX_VALUE;
+        int sum = 0;
+        int maxOfNegative = Integer.MAX_VALUE;
+        int minOfPositive = Integer.MAX_VALUE;
         boolean flag = false;
         
         for(int i=0;i<arr.length;++i){
@@ -21,13 +21,13 @@ class Solution {
         
         if(k%2!=0){
             if(flag){
-                long x = Math.min(maxOfNegative,minOfPositive);
-                return (int)(sum-(2*x));
+                int x = Math.min(maxOfNegative,minOfPositive);
+                return (sum-(2*x));
             }else{
-                return (int)(sum - (2*minOfPositive));
+                return (sum - (2*minOfPositive));
             }
         }
         
-        return (int)sum;
+        return sum;
     }
 }
